Configuration for Azure National Cloud Auth
Adjust configuration of the plugin to allow admin to select which MS Azure cloud to authenticate against. This allows easy selection of the global cloud or a national cloud without requiring a full OAuth URL to be entered. Remove configuration for direct OAuth URL and replace with just needing the AD Tenant ID.